requirements.txt
pyzmq==25.1.2
numpy==1.26.4